Refactor cta in home page
