+++ web/src/pages/home/components/mod.rs
@@ -1,5 +1,7 @@
 pub mod benefits;
+pub mod cta;
 pub mod hero;
 pub mod workflow;
 
+pub use cta::HomePageCta;
 pub use workflow::HomePageWorkflow;
