+++ web/styles/pages/home/_page.scss
@@ -18,82 +18,13 @@
         z-index: 0;
         pointer-events: none;
         filter: blur(10px);
-    }
 
-    > * {
-        position: relative;
-        z-index: 1;
-    }
-
-    &__final-cta {
-        grid-column: 1 / -1;
-        padding: clamp(24px, 3.2vw, 34px);
-        text-align: center;
-
-        h3 {
-            font-size: clamp(28px, 2.9vw, 38px);
-            line-height: 1.15;
-            margin-bottom: 8px;
-        }
-
-        p {
-            font-size: clamp(14px, 1.5vw, 18px);
-            margin-bottom: 18px;
-        }
-
-        .button {
-            min-width: min(320px, 100%);
-        }
-    }
-
-    &__hero,
-    &__final-cta {
-        animation: home-reveal 0.6s ease-out both;
-    }
-
-    &__final-cta {
-        animation-delay: 0.22s;
-    }
-
-    @include media.phone {
-        &::before {
+        @include media.phone {
             inset: -2% 0 auto;
             height: 220px;
         }
 
-        &__hero,
-        &__final-cta {
-            border-radius: 18px;
-        }
-
-        &__logo {
-            align-items: flex-start;
-        }
-
-        &__cta-row {
-            flex-direction: column;
-
-            .button {
-                width: 100%;
-            }
-        }
-
-        &__final-cta {
-            .button {
-                min-width: 100%;
-            }
-        }
-    }
-
-    @media (prefers-reduced-motion: reduce) {
-        &__hero,
-        &__final-cta {
-            animation: none;
-        }
-    }
-
-    @include theme.light {
-        &::before {
+        @include theme.light {
             background: radial-gradient(
                 circle at top center,
                 color.change(colors.$blue-100, $alpha: 0.26),
@@ -101,72 +32,18 @@
             );
         }
 
-        &__hero {
-            &::after {
-                background: linear-gradient(
-                    180deg,
-                    color.change(colors.$white, $alpha: 0.34),
-                    color.change(colors.$white, $alpha: 0)
-                );
-            }
-
-            h1 {
-                color: color.change(colors.$text, $alpha: 0.96);
-            }
-        }
-
-        &__brand-name {
-            color: color.change(colors.$text, $alpha: 0.96);
-        }
-
-        &__eyebrow,
-        &__hero-copy,
-        &__final-cta p {
-            color: color.change(colors.$text, $alpha: 0.8);
-        }
-
-        &__final-cta h3 {
-            color: color.change(colors.$text, $alpha: 0.96);
-        }
-    }
-
-    @include theme.dark {
-        &::before {
+        @include theme.dark {
             background: radial-gradient(
                 circle at top center,
                 color.change(colors.$cyan-100, $alpha: 0.24),
                 color.change(colors.$cyan-100, $alpha: 0) 70%
             );
         }
+    }
 
-        &__hero {
-            &::after {
-                background: linear-gradient(
-                    180deg,
-                    color.change(colors.$background, $alpha: 0.22),
-                    color.change(colors.$background, $alpha: 0)
-                );
-            }
-
-            h1 {
-                color: color.change(colors.$background, $alpha: 0.97);
-            }
-        }
-
-        &__brand-name {
-            color: color.change(colors.$background, $alpha: 0.97);
-        }
-
-        &__eyebrow,
-        &__hero-copy,
-        &__final-cta p {
-            color: color.change(colors.$background, $alpha: 0.82);
-        }
-
-        &__benefits h3,
-        &__final-cta h3 {
-            color: color.change(colors.$background, $alpha: 0.97);
-        }
+    > * {
+        position: relative;
+        z-index: 1;
     }
 }
