+++ web/styles/pages/home/components/index.scss
@@ -1,3 +1,4 @@
 @forward "hero";
 @forward "benefits";
 @forward "workflow";
+@forward "cta";
